rename GA file
